skills/instagram-reels/requirements.txt
playwright>=1.40.0
pillow>=10.0.0
requests>=2.31.0
python-dotenv>=1.0.0
opencv-python>=4.8.0
pytesseract>=0.3.10
colorspacious>=1.1.1
scikit-image>=0.21.0
numpy>=1.24.0
